Add a WPT and update test expectation.
`grid-item-inline-contribution-002.html` fails to properly assert that grid
column sizes are re-resolved in Firefox. For example, using DevTools to inspect
the first column shows it is 10px, yet Firefox still passes the test.

`grid-item-inline-contribution-004.html` is adapted from
`grid-item-inline-contribution-002.html` It correctly tests the re-resolution of
the grid columns sizes by requiring the first column to be exactly 50px in order
to render a green rectangle.