Explicitly return unconstrained size for block masonry in ComputeBSizeForResolvingRowSizes() for legacy track sizing algo.
In legacy track sizing algo, we return unconstrained size for block masonry in
`ComputeBSizeForResolvingRowSizes()` as well, if the logic falls into
`if (!StaticPrefs::layout_css_grid_multi_pass_track_sizing_enabled())`.

This patch is a preparation to move `CalculateTrackSizesForAxis()` call to
`Reflow()` in order to adding extra pass to resolve columns size in bug 1957501.

This patch does not change behavior.