Move CalculateTrackSizesForAxis() from ComputeBSizeForResolvingRowSizes() to Reflow().
`CalculateTrackSizesForAxis()` is the top-level function that implements the
Grid Sizing Algorithm Step 1 to 4. This patch moves the call from
`ComputeBSizeForResolvingRowSizes()` into `Reflow()` to make it clearer that the
track sizing is initiated from there.

This patch prepares for bug 1957501, which will add an additional pass for
resolving the column sizes.

This patch does not change behavior.