Bug 1966450 - Mark grid container as having dirty children if it reflows grid items when computing its intrinsic inline-size.
In Bug 1944423, we introduced logic to resolve row sizes when computing a grid
container's intrinsic inline-size. This can cause grid items to be reflowed to
measure their block-sizes as their content contributions. Since the grid
container is still computing its intrinsic inline-size at this point, the reflow
occurs under provisional inline constraints. As a result, the grid items' sizes
are temporary and must be reflowed again later under the container's final
inline-size.

In `grid-container-as-flex-item-001.html`, when the outer grid container
prepares a `ReflowInput` for the final reflow of its grid item (the flex
container), it calls `GetMinISize()` [1], which recursively calls into the inner
grid container's `GetMinISize()`, triggering the row size resolution and reflows
the inner grid items as described above.

However, if the inner grid container has already been reflowed during a previous
content measurement by the flex container, its block-size might be cached. In
this case, the flex container may skip reflowing the inner grid container,
leading to incorrect layout results as demonstrated in
`grid-container-as-flex-item-001.html`.

This patch fixes the issue by marking the grid container (which is a flex item)
as having dirty children when it reflows grid items during intrinsic size
computation. This informs its parent (the flex container) that the block-size
cache is invalid, and a new measuring reflow or final reflow is needed [2][3].

[1] `GetMinSize()` is called in `nsIFrame::ComputeSize()`:
https://searchfox.org/mozilla-central/rev/fdb34ddfe30bd54aba991feb72b1476c77938e46/layout/generic/nsIFrame.cpp#6771-6773
[2] The flex item's cache is rejected if it has dirty children, triggering a
measuring reflow:
https://searchfox.org/mozilla-central/rev/fdb34ddfe30bd54aba991feb72b1476c77938e46/layout/generic/nsFlexContainerFrame.cpp#2054
[3] If the flex item didn't receive a measuring reflow, it will need a final
reflow:
https://searchfox.org/mozilla-central/rev/fdb34ddfe30bd54aba991feb72b1476c77938e46/layout/generic/nsFlexContainerFrame.cpp#2676-2687

Differential Revision: https://phabricator.services.mozilla.com/D254131