Bug 1977501 Part 2 - Use correct percentage basis when performing measuring reflow for grid items.
`GridReflowInput::PercentageBasisFor()` considers track sizes, and resolve
correct percentage basis for a grid item if a given axis has a definite track
size.

The `aPercentageBasis` argument on `ContentContribution()` is computed in
`CachedIntrinsicSizes` constructor [1].

This patch also updates the documentation for `PercentageBasisFor()`.
Specifically, delete the description of the implementation details that
described the legacy behavior (when
layout.css.grid-multi-pass-track-sizing.enabled=false).

[1] https://searchfox.org/mozilla-central/rev/146c8d8d1f3cf15be3c263808f9674a97752c41a/layout/generic/nsGridContainerFrame.cpp#6001