WIP: Bug 1986837 - patch 3 - When a textrun maps multiple content flows, try to use this to initialize the previous class for autospacing.
This handles cases where there are inline spans with the same font style
(but e.g. different color or text-decoration), so textrun construction
maps multiple elements with a single run for continuous shaping. In this
situation, we can use the mapped-flow data to find the preceding content
and determine the initial prev-class value to use.

This is sufficient to pass the existing text-autospace-elements-002 and -003
testcases, but would not handle cases where there's a textrun-breaking
style change (e.g. font-weight) at the inline boundary.

Differential Revision: https://phabricator.services.mozilla.com/D264614