Bug 1983345 Part 4 - Revise the computation of containing block border-box size.
Previously, when the containing block was in reflow, we get the size from
absolute containing block's reflow input. That works in most cases (because the
two often coincide) but is conceptually wrong when they differ.

This patch computes the border-box size from the information returned from
`GetHypotheticalBoxContainingBlock()` instead, and treat an unconstrained
block-size as 0 for container-size purposes.