Move two abspos reflow methods from nsIFrame to nsContainerFrame.
After changing the table parts to use `nsContainerFrame`, we can move these two
methods into `nsContainerFrame`, and get rid of the type-casting from
`nsIFrame*` to `nsContainerFrame*` in `ReflowAbsoluteFrames`.

Also, revise the documentation for both methods. Note that `nsReflowStatus`
turned into a C++ class that initializes itself, so I delete the description "is
assume to be already-initialized".