Fix a potential writing-mode misuse when initializing abspos child's ReflowInput.
In the old code, `availISize` and `availBSize` are both in the absCB's
writing-mode, but it uses a `LogicalSize` with the kid's writing-mode when
constructing the kid's `ReflowInput`, which doesn't seem correct in a orthogonal
reflow. However, I didn't spend time crafting a testcase, since our abspos
fragmentation is somewhat broken.

Also, with the previous part introducing a logical containing-block
size (`cbSize`) in `ReflowAbsoluteFrame()`. This patch simplifies the code by
removing `logicalCBSize` (which is in the kid's writing-mode).