Update comments in ReflowInput::InitAbsoluteConstraints().
Delete legacy XXX comments. It is true that we always compute the inline-size in
`nsIFrame::ComputeSize()`. However, we compute the offsets in the containing
block's writing mode. In orthogonal flow, the inline-size in containing block's
writing mode is the block-size of `mFrame`, which might still be unconstrained.

Note: changing `iSizeIsAuto` to `computedSize.ISize(cbwm)` could introduce test
failures in
`testing/web-platform/tests/css/css-position/position-absolute-table-001.html`,
because the table wrapper always has `auto` width/height, but
`nsTableWrapperFrame::ComputeAutoSize()` can resolve a definite width/height,
which acts as min-width/min-height. So the offsets are only correct when we
resolve them after reflowing the table.