Bug 198598 - Defer auto margin computation of an abspos frame until it is reflowed.
Remove auto margin resolution in both axes for an abspos frame in
`ReflowInput::InitAbsoluteConstraints()`. Auto margins are now resolved in
`AbsoluteContainingBlock::ResolveAutoMarginsAfterLayout()` after the frame’s
size is known.

Note: the old code in `InitAbsoluteConstraints()` makes the `autoBSize`
non-negative. This patch ports that behavior to make the `stretchFitSize`
non-negative as well, which fixes additional wpts.