Bug 2004166 - Properly make aNewAbsoluteContainingBlock an abspos cb.
If an ib-split `-moz-block-inside-inline-wrapper` is split by a column-span, the
wrapper before column-span might not be an absolute containing block if it does
not have any abspos children. In that case, `nsIFrame::Init()` won't create an
absolute containing block for the continuation.

This patch removes the assertions and properly sets up an absolute containing
block for the frame when it needs one.