Bug 2007602 Part 1 - Fix iterating pushed child list in PrepareAbsoluteFrames(). r?#layout
The old code was wrong to use `GetNextInFlow()` to get the next child in the
loop. It should use `GetNextSibling()`. This patch uses `nsFrameList` iterator,
which hopefully is less error-prone. This fixes a potential bug, but the main
feature is still developing behind bug 1994346, so I didn't try to craft a
testcase.

Differential Revision: https://phabricator.services.mozilla.com/D277493